refactor: moving shared assets to shared folder
