+++ packages/core/components/shared/font-names.ts
@@ -1,0 +1,2 @@
+export const MSSansSerif = 'MS Sans Serif';
+export const R95VideoNumbers = 'React95Video-Numbers';
